+++ Assets/YooAsset/Samples~/Extension Sample/Runtime/ExtensionFileSystem/WechatFileSystem/Operation/WXFSClearUnusedBundleFilesAsync.cs
@@ -1,6 +1,7 @@
 ﻿#if UNITY_WEBGL && WEIXINMINIGAME
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using YooAsset;
 using WeChatWASM;
 
@@ -39,6 +40,7 @@ internal class WXFSClearUnusedBundleFilesAsync : FSClearCacheFilesOperation
         {
             _steps = ESteps.WaitingSearch;
 
+            // 说明：__GAME_FILE_CACHE/yoo/ 目录下包含所有的资源文件和清单文件
             var fileSystemMgr = _fileSystem.GetFileSystemMgr();
             var statOption = new WXStatOption();
             statOption.path = _fileSystem.FileRoot;
@@ -47,13 +49,24 @@ internal class WXFSClearUnusedBundleFilesAsync : FSClearCacheFilesOperation
             {
                 foreach (var fileStat in response.stats)
                 {
-                    // 注意：存储文件必须按照Bundle文件哈希值存储！
-                    string bundleGUID = Path.GetFileNameWithoutExtension(fileStat.path);
+                    // 如果是目录文件
+                    string fileExtension = Path.GetExtension(fileStat.path);
+                    if (string.IsNullOrEmpty(fileExtension))
+                        continue;
+
+                    // 如果是资源清单
+                    //TODO 默认的清单文件格式
+                    if (fileExtension == ".bytes" || fileExtension == ".hash")
+                        continue;
+
+                    // 注意：适配不同的文件命名方式！
+                    string fileNameWithoutExtension = Path.GetFileNameWithoutExtension(fileStat.path);
+                    string bundleGUID = fileNameWithoutExtension.Split('_').Last();
                     if (_manifest.TryGetPackageBundleByBundleGUID(bundleGUID, out PackageBundle value) == false)
                     {
-                        string fullPath = WX.GetCachePath(fileStat.path);
-                        if (_unusedCacheFiles.Contains(fullPath) == false)
-                            _unusedCacheFiles.Add(fullPath);
+                        string filePath = _fileSystem.FileRoot + fileStat.path;
+                        if (_unusedCacheFiles.Contains(filePath) == false)
+                            _unusedCacheFiles.Add(filePath);
                     }
                 }