+++ Assets/YooAsset/Samples~/Extension Sample/Runtime/ExtensionFileSystem/WechatFileSystem/WechatFileSystem.cs
@@ -125,30 +125,30 @@ internal class WechatFileSystem : IFileSystem
         var operation = new WXFSRequestPackageVersionOperation(this, appendTimeTicks, timeout);
         return operation;
     }
-    public virtual FSClearCacheFilesOperation ClearCacheFilesAsync(PackageManifest manifest, string clearMode, object clearParam)
+    public virtual FSClearCacheFilesOperation ClearCacheFilesAsync(PackageManifest manifest, ClearCacheFilesOptions options)
     {
-        if (clearMode == EFileClearMode.ClearAllBundleFiles.ToString())
+        if (options.ClearMode == EFileClearMode.ClearAllBundleFiles.ToString())
         {
             var operation = new WXFSClearAllBundleFilesOperation(this);
             return operation;
         }
-        else if (clearMode == EFileClearMode.ClearUnusedBundleFiles.ToString())
+        else if (options.ClearMode == EFileClearMode.ClearUnusedBundleFiles.ToString())
         {
             var operation = new WXFSClearUnusedBundleFilesAsync(this, manifest);
             return operation;
         }
         else
         {
-            string error = $"Invalid clear mode : {clearMode}";
+            string error = $"Invalid clear mode : {options.ClearMode}";
             var operation = new FSClearCacheFilesCompleteOperation(error);
             return operation;
         }
     }
-    public virtual FSDownloadFileOperation DownloadFileAsync(PackageBundle bundle, DownloadParam param)
+    public virtual FSDownloadFileOperation DownloadFileAsync(PackageBundle bundle, DownloadFileOptions options)
     {
-        param.MainURL = RemoteServices.GetRemoteMainURL(bundle.FileName);
-        param.FallbackURL = RemoteServices.GetRemoteFallbackURL(bundle.FileName);
-        var operation = new WXFSDownloadFileOperation(this, bundle, param);
+        options.MainURL = RemoteServices.GetRemoteMainURL(bundle.FileName);
+        options.FallbackURL = RemoteServices.GetRemoteFallbackURL(bundle.FileName);
+        var operation = new WXFSDownloadFileOperation(this, bundle, options);
         return operation;
     }
     public virtual FSLoadBundleOperation LoadBundleFile(PackageBundle bundle)
@@ -189,11 +189,6 @@ internal class WechatFileSystem : IFileSystem
         if (string.IsNullOrEmpty(_wxCacheRoot))
         {
             throw new System.Exception("请配置微信小游戏缓存根目录！");
-        }
-
-        if (_wxCacheRoot.StartsWith(WX.PluginCachePath) == false)
-        {
-            _wxCacheRoot = PathUtility.Combine(WX.PluginCachePath, _wxCacheRoot);
         }
 
         // 注意：CDN服务未启用的情况下，使用微信WEB服务器